Major architectural changes
Implemented the dual DOM model! The computation and manipulation of DOM
elements has been split from the actual handling of those elements in
the view hierarchy resulting in two DOMs that I dub the static DOM and
the dynamic DOM respectively.

Added preliminary support for a special "contents" attribute. This
special attribute behaves for the most part as a normal attribute, but
allows for the referencing and overriding of the actual contents of a
node.